metal: refactor shared textures
